refactored helpers for icons, seperated pages and components, seperated styled components
