+++ src/Helpers/Icons.js
@@ -1,0 +1,63 @@
+// Home Page
+import LikeIcon from "../assets/img/Heart.png";
+import MessageIcon from "../assets/img/Message.png";
+
+// Top Bar
+import CameraIcon from "../assets/img/Camera.png";
+import InstagramTitle from "../assets/img/Instagram2.png";
+
+// Stories
+import StoryYiqi from "../assets/img/StoryYiqi.png";
+import StoryInstagram from "../assets/img/StoryInsta.png";
+import StoryCB from "../assets/img/StoryCB.png";
+import StoryBanksy from "../assets/img/StoryBanksy.png";
+import StoryLR from "../assets/img/StoryLR.png";
+import StoryEmilie from "../assets/img/StoryEmilie.png";
+
+// Posts
+import PostImageOne from "../assets/img/PostOne.JPG";
+import PostImageTwo from "../assets/img/PostTwo.JPG";
+import PostImageThree from "../assets/img/PostThree.JPG";
+import ProfilePic from "../assets/img/StoryYiqi.png";
+import CommentIcon from "../assets/img/Comment.png";
+import SaveIcon from "../assets/img/Save.png";
+
+// Bottom Bar
+import HomeIcon from "../assets/img/Home.png";
+import SearchIcon from "../assets/img/Search.png";
+import AddIcon from "../assets/img/Add.png";
+import ProfileIcon from "../assets/img/StoryYiqi.png";
+
+//Top Bar
+export const HeaderIcons = [CameraIcon, InstagramTitle, MessageIcon];
+
+// Stories
+export const StoryIcons = [
+  StoryYiqi,
+  StoryCB,
+  StoryInstagram,
+  StoryEmilie,
+  StoryBanksy,
+  StoryLR,
+];
+
+// Posts
+export const PostIcons = [
+  PostImageOne,
+  PostImageTwo,
+  PostImageThree,
+  ProfilePic,
+  LikeIcon,
+  CommentIcon,
+  MessageIcon,
+  SaveIcon,
+];
+
+// Bottom Bar
+export const BottomIcons = [
+  HomeIcon,
+  SearchIcon,
+  AddIcon,
+  LikeIcon,
+  ProfileIcon,
+];
