+++ src/components/Home/Bottom/BottomBar.js
@@ -1,20 +1,17 @@
 import React from "react";
 
-import HomeIcon from "../../assets/img/Home.png";
-import SearchIcon from "../../assets/img/Search.png";
-import LikeIcon from "../../assets/img/Heart.png";
-import AddIcon from "../../assets/img/Add.png";
-import ProfileIcon from "../../assets/img/StoryYiqi.png";
+//Helpers
+import { BottomIcons } from "../../../Helpers/Icons";
 
+//Styles
 import {
   BottomContainerDiv,
   BottomIconDiv,
   BottomIconImage,
   IconContainerDiv,
-} from "./Home.styles";
+} from "./BottomBar.styles.js";
 
 const BottomBar = () => {
-  const BottomIcons = [HomeIcon, SearchIcon, AddIcon, LikeIcon, ProfileIcon];
   return (
     <BottomContainerDiv>
       <BottomIconDiv>
