+++ src/components/Home/Bottom/BottomBar.styles.js
@@ -1,0 +1,31 @@
+import styled from "styled-components";
+
+export const BottomContainerDiv = styled.div`
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  bottom: 0;
+  width: 100vw;
+  height: 3rem;
+  background: #252525;
+  color: white;
+`;
+
+export const BottomIconDiv = styled.div`
+  display: flex;
+  justify-content: space-around;
+  width: 100vw;
+`;
+
+export const BottomIconImage = styled.img`
+  max-width: 3rem;
+  height: 60%;
+  padding: 0.3em;
+`;
+
+export const IconContainerDiv = styled.div`
+  width: 25%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
