+++ src/components/Home/Header/HeaderBar.js
@@ -1,28 +1,28 @@
 import React from "react";
 
+// Helpers
+import { HeaderIcons } from "../../../Helpers/Icons";
+
+// Styles
 import {
   HeaderContainerDiv,
   HeaderIconDiv,
   HeaderIconImage,
-} from "./Home.styles";
-
-import CameraIcon from "../../assets/img/Camera.png";
-import MessageIcon from "../../assets/img/Message.png";
-import InstagramTitle from "../../assets/img/Instagram2.png";
+} from "./HeaderBar.styles.js";
 
 const HeaderBar = () => {
   return (
     <HeaderContainerDiv>
       <HeaderIconDiv>
-        <HeaderIconImage src={CameraIcon} alt="" />
+        <HeaderIconImage src={HeaderIcons[0]} alt="" />
         <HeaderIconImage
           style={{ maxWidth: "none" }}
-          src={InstagramTitle}
+          src={HeaderIcons[1]}
           alt=""
         />
         <HeaderIconImage
           style={{ padding: "0.4em 0.3em 0 0" }}
-          src={MessageIcon}
+          src={HeaderIcons[2]}
           alt=""
         />
       </HeaderIconDiv>
