+++ src/components/Home/Header/HeaderBar.styles.js
@@ -1,0 +1,24 @@
+import styled from "styled-components";
+
+export const HeaderContainerDiv = styled.div`
+  display: flex;
+  justify-content: center;
+  position: fixed;
+  top: 0;
+  width: 100vw;
+  height: 3rem;
+  background: #252525;
+  color: white;
+`;
+
+export const HeaderIconDiv = styled.div`
+  display: flex;
+  justify-content: space-between;
+  width: 100vw;
+`;
+
+export const HeaderIconImage = styled.img`
+  max-width: 3rem;
+  height: 80%;
+  padding: 0.4em 0 0 0.3em;
+`;
